test fixes

+++ Storage/ElasticaStorage.php
@@ -10,14 +10,17 @@ namespace Phlexible\Bundle\IndexerStorageElasticaBundle\Storage;
 
 use Elastica\Client;
 use Elastica\Filter\Ids;
+use Elastica\Filter\MatchAll;
 use Elastica\Index;
 use Elastica\Query as ElasticaQuery;
 use Elastica\Query;
 use Elastica\Result;
+use Elastica\ResultSet;
 use Phlexible\Bundle\IndexerBundle\Document\DocumentInterface;
 use Phlexible\Bundle\IndexerBundle\Storage\Flushable;
 use Phlexible\Bundle\IndexerBundle\Storage\Optimizable;
 use Phlexible\Bundle\IndexerBundle\Storage\StorageInterface;
+use Phlexible\Bundle\IndexerBundle\Storage\UpdateQuery\Command\CommandCollection;
 use Phlexible\Bundle\IndexerBundle\Storage\UpdateQuery\UpdateQuery;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
@@ -39,6 +42,11 @@ class ElasticaStorage implements StorageInterface, Optimizable, Flushable
     private $mapper;
 
     /**
+     * @var UpdateQuery
+     */
+    private $updateQuery;
+
+    /**
      * @var EventDispatcherInterface
      */
     private $eventDispatcher;
@@ -51,13 +59,20 @@ class ElasticaStorage implements StorageInterface, Optimizable, Flushable
     /**
      * @param Client                   $client
      * @param ElasticaMapper           $mapper
+     * @param UpdateQuery              $updateQuery
      * @param EventDispatcherInterface $eventDispatcher
      * @param string                   $indexName
      */
-    public function __construct(Client $client, ElasticaMapper $mapper, EventDispatcherInterface $eventDispatcher, $indexName)
+    public function __construct(
+        Client $client,
+        ElasticaMapper $mapper,
+        UpdateQuery $updateQuery,
+        EventDispatcherInterface $eventDispatcher,
+        $indexName)
     {
         $this->client = $client;
         $this->mapper = $mapper;
+        $this->updateQuery = $updateQuery;
         $this->eventDispatcher = $eventDispatcher;
         $this->indexName = $indexName;
     }
@@ -65,17 +80,25 @@ class ElasticaStorage implements StorageInterface, Optimizable, Flushable
     /**
      * {@inheritdoc}
      */
-    public function createUpdate()
+    public function createCommands()
     {
-        return new UpdateQuery($this->eventDispatcher);
+        return $this->updateQuery->createCommands();
     }
 
     /**
      * {@inheritdoc}
      */
-    public function execute(UpdateQuery $update)
+    public function runCommands(CommandCollection $commands)
     {
-        return $update->execute($this);
+        return $this->updateQuery->run($this, $commands);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function queueCommands(CommandCollection $commands)
+    {
+        return $this->updateQuery->queue($commands);
     }
 
     /**
@@ -144,6 +167,10 @@ class ElasticaStorage implements StorageInterface, Optimizable, Flushable
     {
         $result = $this->find($identifier);
 
+        if (!$result) {
+            return;
+        }
+
         $this->client->deleteIds(array($identifier), $this->getIndexName(), $result->getType());
     }
 
@@ -152,7 +179,20 @@ class ElasticaStorage implements StorageInterface, Optimizable, Flushable
      */
     public function deleteAll()
     {
-        $this->getIndex()->deleteDocuments(array());
+        $resultSet = $this->findAll();
+
+        if (!$resultSet) {
+            return;
+        }
+
+        $typeIds = array();
+        foreach ($resultSet->getResults() as $result) {
+            $typeIds[$result->getType()][] = $result->getId();
+        }
+
+        foreach ($typeIds as $type => $ids) {
+            $this->client->deleteIds($ids, $this->getIndexName(), $type);
+        }
     }
 
     /**
@@ -211,5 +251,18 @@ class ElasticaStorage implements StorageInterface, Optimizable, Flushable
         $resultSet = $this->getIndex()->search($query);
 
         return $resultSet->current();
+    }
+
+    /**
+     * @return ResultSet
+     */
+    private function findAll()
+    {
+        $query = new ElasticaQuery();
+        $filter = new MatchAll();
+        $query->setPostFilter($filter);
+        $resultSet = $this->getIndex()->search($query);
+
+        return $resultSet;
     }
 }
